requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.6
django-redis==5.4.0
django-ratelimit==4.1.0
python-decouple==3.8
gunicorn==21.2.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
